All simultaneous recordings on the same frequency
On some single source tuners you can play several channels at the same time if
they are on the same frequency/mux/transponder. This commit makes sure that
recordings on different channels can overlap, if they are on the same frequency.

This feature is optional and has to enabled in config.toml.